Fix/update sent location (#10)
* Add LICENSE.md.

* Update sentiment dictionary location in README.md.

* Refactor and enhance error handling

- Refactored `uploadCommand` function to include try-catch block for improved error handling.
- Added error logging using Timber for better debugging.
- Updated `Instrumentation.java` to handle UiAutomation connection failures gracefully.
- Improved resource management and cleanup in `uploadCommand`.

+++ app/src/androidTest/java/com/screenlake/data/TestUploadHandler.kt
@@ -26,16 +26,20 @@ class TestUploadHandler(
         test: Boolean,
         testContext: Context?
     ) {
-        // Simulate a successful upload
-        Timber.tag(TAG).d("Simulating upload of file ${file.name}")
+        if (isNetworkConnected) {
+            // Simulate a successful upload
+            Timber.tag(TAG).d("Simulating upload of file ${file.name}")
 
-        // Simulate upload delay
-        delay(1000)
+            // Simulate upload delay
+            delay(1000)
 
-        file.delete()
-        entryId?.let { generalOperationsRepository.deleteZip(it) }
-        UploadWorker.uploadFeedback.postValue("Upload succeeded -> Simulated path")
-        Timber.tag(TAG).d("Upload simulated as successful")
+            file.delete()
+            entryId?.let { generalOperationsRepository.deleteZip(it) }
+            UploadWorker.uploadFeedback.postValue("Upload succeeded -> Simulated path")
+            Timber.tag(TAG).d("Upload simulated as successful")
+        }else{
+
+        }
     }
 
     override suspend fun isNetworkConnected(): Boolean {
